+++ apps/web/client/src/server/api/root.ts
@@ -1,5 +1,5 @@
 import { createCallerFactory, createTRPCRouter } from '~/server/api/trpc';
-import { canvasRouter, chatRouter, codeRouter, domainRouter, frameRouter, githubRouter, invitationRouter, memberRouter, projectRouter, sandboxRouter, subscriptionRouter, userCanvasRouter, userRouter } from './routers';
+import { canvasRouter, chatRouter, codeRouter, domainRouter, frameRouter, githubRouter, invitationRouter, memberRouter, projectRouter, sandboxRouter, subscriptionRouter, userCanvasRouter, userRouter, usageRouter } from './routers';
 
 /**
  * This is the primary router for your server.
@@ -20,6 +20,7 @@ export const appRouter = createTRPCRouter({
     domain: domainRouter,
     github: githubRouter,
     subscription: subscriptionRouter,
+    usage: usageRouter,
 });
 
 // export type definition of API
