+++ packages/db/src/schema/subscription/usage.ts
@@ -1,14 +1,12 @@
 import { relations } from 'drizzle-orm';
 import { pgTable, text, timestamp, uuid } from 'drizzle-orm/pg-core';
 import { users } from '../user';
-import { subscriptions } from './subscription';
 
 export const usageRecords = pgTable('usage_records', {
     id: uuid('id').defaultRandom().primaryKey(),
 
     // Relationships
     userId: uuid('user_id').notNull().references(() => users.id),
-    subscriptionId: uuid('subscription_id').notNull().references(() => subscriptions.id),
 
     // Metadata
     type: text('type', { enum: ['message'] }).notNull(),
@@ -16,9 +14,9 @@ export const usageRecords = pgTable('usage_records', {
 }).enableRLS();
 
 export const usageRelations = relations(usageRecords, ({ one }) => ({
-    subscription: one(subscriptions, {
-        fields: [usageRecords.subscriptionId],
-        references: [subscriptions.id],
+    user: one(users, {
+        fields: [usageRecords.userId],
+        references: [users.id],
     })
 }))
